+++ src/main/java/com/homework/homework_week2/post/dto/PostResponseDto.java
@@ -13,7 +13,7 @@ public class PostResponseDto {
     private Long postId;
     private String title;
     private String content;
-    private Long template;
+    private int template;
     private String imageUrl;
     private Long viewCount;
     private List<CommentResponseDto> comments;
